requirements.txt
altair==5.5.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
filelock==3.17.0
fonttools==4.55.8
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.24.1
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
plotly-express==0.4.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
